Enable/Disable the (systemd) `docker.service` and `.socket`
When using this role to install `docker` but only intend for it's use in 
`rootless` mode, the Docker documentation suggests disabling 
`docker.service`.

However, that also leaves the `docker.socket` running, which whenever it 
receieves a connect, the daemon (service) is restarted.  

This patch enables/disables the both the service and the socket.